Remove configuration file option
This program originally was created with an option for a configuration
file (~/.config/cdhist-flags.conf) and was later augmented with the
ability to add startup options when defining the alias. This second
approach is better than the first (because you can have different
options for different aliases) so this commit removes that first option.

I have seen that cdhist can be a bit laggy on under-resourced machines
like the Raspberry Pi so, in addition to code simplification, this
change also (very slightly?) improves performance because it removes a
couple of large package imports and avoids reading the file system every
run looking for that config file.

I also removed some code added at version 2 that checked and accommodated
a very old legacy --shell option.